+++ packages/kit/test/ambient.d.ts
@@ -16,7 +16,7 @@ declare global {
 	const preloadData: (url: string) => Promise<void>;
 	const beforeNavigate: (fn: (url: URL) => void | boolean) => void;
 	const afterNavigate: (fn: () => void) => void;
-	const preloadCode: (...urls: string[]) => Promise<void>;
+	const preloadCode: (pathname: string) => Promise<void>;
 }
 
 export {};