+++ packages/kit/test/apps/basics/src/routes/navigation-lifecycle/after-navigate-properly-removed/b/+page.svelte
@@ -1,0 +1,7 @@
+<h1>/B</h1>
+<div>
+	was called:
+	<span class="nav-lifecycle-after-nav-removed-test-target">false</span>
+</div>
+<a href="/navigation-lifecycle/after-navigate-properly-removed/a">/a</a>
+<a href="/navigation-lifecycle/after-navigate-properly-removed/b">/b</a>