+++ packages/kit/test/apps/options/source/pages/csp-with-stream/+page.server.js
@@ -1,0 +1,5 @@
+export function load() {
+	return {
+		lazy: new Promise((resolve) => setTimeout(() => resolve(), 1000)).then(() => 'Moo Deng!')
+	};
+}