+++ packages/kit/test/build-errors/prerender.spec.js
@@ -4,26 +4,28 @@ import path from 'node:path';
 import { EOL } from 'node:os';
 import process from 'node:process';
 
-test('prerenderable routes must be prerendered', () => {
+const timeout = 60_000;
+
+test('prerenderable routes must be prerendered', { timeout }, () => {
 	assert.throws(
 		() =>
 			execSync('pnpm build', {
 				cwd: path.join(process.cwd(), 'apps/prerenderable-not-prerendered'),
 				stdio: 'pipe',
-				timeout: 60000
+				timeout
 			}),
 		/The following routes were marked as prerenderable, but were not prerendered because they were not found while crawling your app:\r?\n {2}- \/\[x\]/gs
 	);
 });
 
-test('entry generators should match their own route', () => {
+test('entry generators should match their own route', { timeout }, () => {
 	assert.throws(
 		() =>
 			execSync('pnpm build', {
 				cwd: path.join(process.cwd(), 'apps/prerender-entry-generator-mismatch'),
 				stdio: 'pipe',
-				timeout: 60000
+				timeout
 			}),
-		`Error: The entries export from /[slug]/[notSpecific] generated entry /whatever/specific, which was matched by /[slug]/specific - see the \`handleEntryGeneratorMismatch\` option in https://kit.svelte.dev/docs/configuration#prerender for more info.${EOL}To suppress or handle this error, implement \`handleEntryGeneratorMismatch\` in https://kit.svelte.dev/docs/configuration#prerender`
+		`Error: The entries export from /[slug]/[notSpecific] generated entry /whatever/specific, which was matched by /[slug]/specific - see the \`handleEntryGeneratorMismatch\` option in https://svelte.dev/docs/kit/configuration#prerender for more info.${EOL}To suppress or handle this error, implement \`handleEntryGeneratorMismatch\` in https://svelte.dev/docs/kit/configuration#prerender`
 	);
 });